Fix weird error introduced by Xcode 10.2 migcom
This version of mig adds mach_msg_destroy_from_kernel()
calls to every generated mig stub. This function does
not appear to be defined anywhere (not in xnu, and not
in libsyscall). As such, the stubs will not link as-is.

I added a patch which creates a dummy definition of
this function. More investigation into the root cause
is called for; however, the version of mig used in Xcode
10.2 does not appear to be open-source, and no other
versions of mig include this code. How strange.